# openapi.yaml
openapi: 3.0.3

info:
  title: Cairn API
  description: |
    A service catalog with a maturity scorecard. Register services, evaluate
    them against a weighted set of checks derived from their GitHub
    repository, and read back a 0-100 score and a Bronze/Silver/Gold level.
  version: "1.0.0"
  license:
    name: MIT
    url: https://github.com/cpkeller25/cairn/blob/main/LICENSE

servers:
  - url: http://localhost:8081
    description: Local development

tags:
  - name: services
    description: The service catalog
  - name: checks
    description: Scorecard rule definitions
  - name: operations
    description: Health, readiness and metrics

paths:
  /api/v1/services:
    post:
      tags: [services]
      summary: Register a service
      operationId: createService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URL of the created service
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Service' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: A service with that name already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          $ref: '#/components/responses/ValidationFailed'

    get:
      tags: [services]
      summary: List the catalog with each service's latest score
      operationId: listServices
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [services]
                properties:
                  services:
                    type: array
                    items: { $ref: '#/components/schemas/Service' }

  /api/v1/services/{id}:
    parameters:
      - $ref: '#/components/parameters/ServiceId'

    get:
      tags: [services]
      summary: Get one service with its scorecard breakdown
      operationId: getService
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Service' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [services]
      summary: Remove a service and its scorecards
      operationId: deleteService
      responses:
        '204':
          description: Deleted, no content
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/services/{id}/evaluate:
    parameters:
      - $ref: '#/components/parameters/ServiceId'

    post:
      tags: [services]
      summary: Gather repository facts, run the scorecard, and persist the result
      operationId: evaluateService
      description: |
        Reads the service's repository through the GitHub API, runs every
        check, stores the result, and returns the service with its new
        scorecard. Previous scorecards are retained as history.
      responses:
        '200':
          description: Evaluated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Service' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: |
            The repository could not be read: not a github.com URL, malformed,
            nonexistent, or private without a configured token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '502':
          description: GitHub was unreachable or rate limited
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/v1/checks:
    get:
      tags: [checks]
      summary: List check definitions and their weights
      operationId: listChecks
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [checks]
                properties:
                  checks:
                    type: array
                    items: { $ref: '#/components/schemas/CheckDefinition' }

  /healthz:
    get:
      tags: [operations]
      summary: Liveness probe
      description: |
        Reports that the process is running. Touches no dependencies, so a
        database outage cannot trigger a restart loop.
      operationId: healthz
      responses:
        '200':
          description: Alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }

  /readyz:
    get:
      tags: [operations]
      summary: Readiness probe
      description: Reports whether the database is reachable.
      operationId: readyz
      responses:
        '200':
          description: Ready to serve traffic
        '503':
          description: Not ready; the database is unreachable

  /metrics:
    get:
      tags: [operations]
      summary: Prometheus metrics
      operationId: metrics
      responses:
        '200':
          description: Metrics in Prometheus text exposition format
          content:
            text/plain:
              schema: { type: string }

components:
  parameters:
    ServiceId:
      name: id
      in: path
      required: true
      description: Service UUID
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: Malformed request or a non-UUID id
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No service with that id
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: The request parsed, but its contents violate the rules
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }

  schemas:
    CreateServiceRequest:
      type: object
      required: [name, owner_team, repo_url]
      properties:
        name:
          type: string
          maxLength: 100
          pattern: '^[a-z0-9]([a-z0-9-]*[a-z0-9])?$'
          description: Lowercase alphanumeric, optionally hyphen-separated
          example: payments-api
        description:
          type: string
          maxLength: 500
          example: Handles payment processing
        owner_team:
          type: string
          maxLength: 100
          example: checkout
        repo_url:
          type: string
          description: GitHub repository URL. The scheme may be omitted.
          example: https://github.com/acme/payments-api
        tier:
          type: integer
          minimum: 1
          maximum: 3
          default: 3
          description: 1 is most critical
          example: 1

    Service:
      type: object
      required: [id, name, owner_team, repo_url, tier, created_at, updated_at, scorecard]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: payments-api }
        description: { type: string }
        owner_team: { type: string, example: checkout }
        repo_url: { type: string }
        tier: { type: integer, minimum: 1, maximum: 3 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        scorecard:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Scorecard'
          description: null until the service has been evaluated

    Scorecard:
      type: object
      required: [overall_score, level, checks]
      properties:
        overall_score:
          type: integer
          minimum: 0
          maximum: 100
          example: 100
        level:
          type: string
          enum: [bronze, silver, gold]
          example: gold
        checks:
          type: array
          items: { $ref: '#/components/schemas/CheckResult' }

    CheckResult:
      type: object
      required: [key, passed, weight, detail]
      properties:
        key: { type: string, example: has_readme }
        passed: { type: boolean, example: true }
        weight: { type: integer, example: 10 }
        detail:
          type: string
          description: Human-readable explanation of the outcome
          example: README found

    CheckDefinition:
      type: object
      required: [key, description, weight]
      properties:
        key: { type: string, example: has_ci }
        description: { type: string, example: Repository has a CI pipeline }
        weight: { type: integer, example: 20 }

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: service not found

    ValidationError:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: validation failed
        fields:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string, example: name }
              message:
                type: string
                example: must be lowercase alphanumeric, optionally hyphen-separated